+++ tests/testthat/test_init_phg.R
@@ -1,0 +1,6 @@
+test_that("PHG initialization", {
+    expect_error(initPhg("dir"))
+    expect_error(initPhg(mtcars))
+})
+
+